+++ RELEASES.md
@@ -1519,3 +1519,13 @@ Based on:
 - [typescript v0.7.0-alpha.20] .
 ### Releases
 - [NPM v0.7.0-alpha.20] https://www.npmjs.com/package/cribl-control-plane/v/0.7.0-alpha.20 - .
+
+## 2026-04-16 02:42:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.5 (2.879.13) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.7.0-alpha.21] .
+### Releases
+- [NPM v0.7.0-alpha.21] https://www.npmjs.com/package/cribl-control-plane/v/0.7.0-alpha.21 - .